chore:book

+++ database/migrations/2023_11_30_133746_create_initialamount_table.php
@@ -10,10 +10,11 @@ return new class extends Migration {
      *
      * @return void
      */
-    public function up() {
+    public function up()
+    {
         Schema::create('initialamount', function (Blueprint $table) {
             $table->id();
-            $table->decimal('amount', 40, 2);
+            $table->double('amount');
             $table->date('createDate');
             $table->string('teamId');
             $table->timestamps();
@@ -25,7 +26,8 @@ return new class extends Migration {
      *
      * @return void
      */
-    public function down() {
+    public function down()
+    {
         Schema::dropIfExists('initialamount');
     }
 };
